--- CUnit/0hana.C
@@ -1,1 +1,0 @@
-../0hana.C
